binance/requirements.txt
polars
duckdb
pyarrow
httpx
mlflow
rich
typer
jinja2
lean
matplotlib
pyyaml
